jiggle 2: the rejiggling
